randomGenerator funcional y modificacion de length en validaciones de product
