+++ api/src/routes/index.js
@@ -9,7 +9,8 @@ const favoriteRouter = require("./favoriteRouter");
 // Ejemplo: const authRouter = require('./auth.js');
 
 const router = Router();
-
+const { generator } = require("../randomGenerator");
+generator();
 router.use("/user", userRoutes);
 router.use("/products", productsRoutes);
 router.use("/invoices", invoicesRoutes);
